requirements.txt
pandas
spotipy
pygsheets
python-telegram-bot